+++ src/qtbox/progressbar/style/demo8.py
@@ -33,7 +33,7 @@ class QtBoxStyleProgressBar8(QProgressBar):
         arc_pen.setColor(QColor("#f1964c"))
         painter.setPen(arc_pen)
         percent = self.value() / self.maximum() if self.maximum() != 0 else 0
-        painter.drawArc(arc_rect, 90*16, percent*360*16)
+        painter.drawArc(arc_rect, 90*16, int(percent*360*16))
 
         font = QFont()
         font.setPixelSize(18)
@@ -75,7 +75,7 @@ class QtBoxStyleProgressBar8(QProgressBar):
 #         arc_pen.setColor(QColor("#f1964c"))
 #         painter.setPen(arc_pen)
 #         percent = self.value() / self.maximum() if self.maximum() != 0 else 0
-#         painter.drawArc(arc_rect, 90*16, percent*360*16)
+#         painter.drawArc(arc_rect, 90*16, int(percent*360*16))
 
 #         font = QFont()
 #         font.setPixelSize(18)
